# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSE341 Project 2 API",
        "version": "1.0.0",
        "description": "API for managing contacts and user profiles with full CRUD operations.",
        "contact": {
            "name": "CSE341 Course",
            "url": "https://www.byupattheway.org"
        },
        "license": {
            "name": "ISC"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development Server"
        },
        {
            "url": "https://your-render-app.onrender.com",
            "description": "Production Server (Render)"
        }
    ],
    "paths": {
...
        "/api/profiles": {
            "get": {
                "summary": "Retrieve all profiles",
                "description": "Fetches a list of all user profiles in the database",
                "operationId": "getAllProfiles",
                "tags": ["Profiles"],
                "responses": {
                    "200": {
                        "description": "A list of profiles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Profile" }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error retrieving profiles",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new profile",
                "description": "Creates a new user profile with all required fields. Profiles have 8 fields: name, email, phone, bio, skills, experience, education, and location.",
                "operationId": "createProfile",
                "tags": ["Profiles"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ProfileInput" }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } }
                        }
                    },
                    "500": {
                        "description": "Server error creating profile",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
                        }
                    }
                }
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "summary": "Retrieve a profile by ID",
                "description": "Fetches a single user profile by its MongoDB _id",
                "operationId": "getProfileById",
                "tags": ["Profiles"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": {
                        "description": "Profile found",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } }
                    },
                    "400": { "description": "Invalid ID format" },
                    "404": { "description": "Profile not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "summary": "Update a profile by ID",
                "description": "Updates one or more fields of an existing profile",
                "operationId": "updateProfile",
                "tags": ["Profiles"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "email": { "type": "string", "format": "email" },
                                    "phone": { "type": "string" },
                                    "bio": { "type": "string" },
                                    "skills": { "type": "array", "items": { "type": "string" } },
                                    "experience": { "type": "string" },
                                    "education": { "type": "string" },
                                    "location": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } }
                    },
                    "400": { "description": "Validation error or invalid ID" },
                    "404": { "description": "Profile not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "summary": "Delete a profile by ID",
                "description": "Permanently removes a profile from the database",
                "operationId": "deleteProfile",
                "tags": ["Profiles"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": {
                        "description": "Profile deleted",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
                            }
                        }
                    },
                    "400": { "description": "Invalid ID format" },
                    "404": { "description": "Profile not found" },
                    "500": { "description": "Server error" }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "type": "object",
                "required": ["name", "email", "phone", "bio", "skills", "experience", "education", "location"],
                "properties": {
                    "_id": { "type": "string", "description": "MongoDB ObjectId" },
                    "name": { "type": "string", "description": "Full name" },
                    "email": { "type": "string", "format": "email", "description": "Email address" },
                    "phone": { "type": "string", "description": "Phone number" },
                    "bio": { "type": "string", "description": "Short biography" },
                    "skills": { "type": "array", "items": { "type": "string" }, "description": "List of skills" },
                    "experience": { "type": "string", "description": "Work experience summary" },
                    "education": { "type": "string", "description": "Education background" },
                    "location": { "type": "string", "description": "Current location" },
                    "createdAt": { "type": "string", "format": "date-time", "description": "Creation timestamp" }
                }
            },
            "ProfileInput": {
                "type": "object",
                "required": ["name", "email", "phone", "bio", "skills", "experience", "education", "location"],
                "properties": {
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "phone": { "type": "string" },
                    "bio": { "type": "string" },
                    "skills": { "type": "array", "items": { "type": "string" } },
                    "experience": { "type": "string" },
                    "education": { "type": "string" },
                    "location": { "type": "string" }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" },
                    "error": { "type": "string" }
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" },
                    "errors": { "type": "array", "items": { "type": "string" } }
                }
            }
        }
    },
...
}